requirements.txt
streamlit==1.31.1
pandas==2.2.0
numpy
groq==0.4.1
python-dotenv==1.0.0
matplotlib
seaborn
google-generativeai
scikit-learn>=1.0.2
